requirements.txt
python-whois==0.8.0
tqdm==4.66.1 